Task 06 from 'Transact SQL' is solved.

+++ Tasks/Transact SQL/01.3.CreateTableAccounts.sql
@@ -5,5 +5,7 @@ Create Table Accounts
 (
 	Id int PRIMARY KEY,
 	PersonId int FOREIGN KEY REFERENCES Persons(Id),
-	Balance float
+	Balance money
 )
+
+
