+++ Tasks/Transact SQL/02.CreateStoredProcedureWithParam.sql
@@ -2,11 +2,11 @@ USE Bank
 GO
 
 CREATE PROCEDURE SelectPersons
-	@Ballance float
-AS
-SELECT Persons.*
-FROM Persons
-INNER JOIN Accounts
-ON Persons.Id = Accounts.PersonId
-WHERE Accounts.Balance > @Ballance
+	@Ballance money
+	AS
+	SELECT Persons.*
+	FROM Persons
+	INNER JOIN Accounts
+	ON Persons.Id = Accounts.PersonId
+	WHERE Accounts.Balance > @Ballance
 GO
